+++ src/ideabox/policy/content/priority_action.py
@@ -34,7 +34,10 @@ class PriorityAction(Project):
 class PriorityActionView(ProjectView):
     def get_state_progress(self):
         return api.content.find(
-            context=self.context, portal_type="state_progress", sort_on="state_date"
+            context=self.context,
+            portal_type="state_progress",
+            sort_on="state_date",
+            sort_order="descending",
         )[:4]
 
     def format_date(self, value):